Updated CredentialOfferEndpoint

+++ app/src/main/java/com/credman/cmwallet/openid4vci/data/TokenEndpoint.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 data class TokenResponse (
     @SerialName("access_token") val accessToken: String,
     @SerialName("refresh_token") val refreshToken: String? = null,
-    @SerialName("expires_in") val expiresInSeconds: Int? = null,
+    @SerialName("expires_in") val expiresInSeconds: Long? = null,
     @SerialName("token_type") val tokenType: String? = null,
     @SerialName("scope") val scopes: List<String>? = null,
 )
